Fix Huffman decoding in cache manager

+++ id_vh.h
@@ -33,7 +33,7 @@ typedef	struct
 
 typedef struct
 {
-	int width,height;
+	short width,height;
 } pictabletype;
